pyproject.toml
[project]
name = "warp_hint"
version = "0.1.0"
description = "Type enhanced Nvidia Warp"
authors = [{ name = "whitingyan", email = "1712428442@qq.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
dependencies = ["warp-lang>=1.8.1", "numpy>=2.3.2"]


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "-e file:///${PROJECT_ROOT}/../warp#egg=warp-lang",
    "-e file:///${PROJECT_ROOT}/#egg=warp-hint",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
]
